Add option to toggle info labels
This adds an option to toggle showing the 'Project' and 'Plan' labels in
the Finger response. This is exposed via an environment variable called
`FINGER_INFO_LABELS` and is enabled by default, preserving the current
behavior.

This allows for a little more customization in the Finger output, albeit
potentially less 'standard'.